[Spaces] Create Default Space on startup (#19177)
When Kibana starts, a Default Space will be created if one does not already exist. The Default Space will contain all objects that are not currently assigned to any other space (i.e., objects created prior to Spaces being enabled, or objects created directly within the Default Space).

To enable this, this PR also introduces the concept of Reserved Spaces, which is denoted by a `_reserved` property on the `space` object. Reserved Spaces:
1. Cannot be deleted
2. Cannot have their URL Context Changed

This PR does not address:
1. Disabling the UI for reserved roles - this will be enabled via https://github.com/elastic/kibana/pull/19035 once this PR is merged into it.

+++ x-pack/plugins/spaces/server/routes/api/v1/spaces.js
@@ -10,6 +10,7 @@ import { getClient } from '../../../../../../server/lib/get_client_shield';
 import { routePreCheckLicense } from '../../../lib/route_pre_check_license';
 import { spaceSchema } from '../../../lib/space_schema';
 import { wrapError } from '../../../lib/errors';
+import { isReservedSpace } from '../../../../common/is_reserved_space';
 import { createDuplicateContextQuery } from '../../../lib/check_duplicate_context';
 
 export function initSpacesApi(server) {
@@ -104,7 +105,7 @@ export function initSpacesApi(server) {
         overwrite = false
       } = request.query;
 
-      const space = omit(request.payload, ['id']);
+      const space = omit(request.payload, ['id', '_reserved']);
 
       const { error } = await checkForDuplicateContext(space);
 
@@ -116,8 +117,16 @@ export function initSpacesApi(server) {
 
       let result;
       try {
+        const existingSpace = await getSpaceById(client, id);
+
+        // Reserved Spaces cannot have their _reserved or urlContext properties altered.
+        if (isReservedSpace(existingSpace)) {
+          space._reserved = true;
+          space.urlContext = existingSpace.urlContext;
+        }
+
         result = await client.create('space', { ...space }, { id, overwrite });
-      } catch(e) {
+      } catch (e) {
         return reply(wrapError(e));
       }
 
@@ -172,8 +181,13 @@ export function initSpacesApi(server) {
       let result;
 
       try {
+        const existingSpace = await getSpaceById(client, id);
+        if (isReservedSpace(existingSpace)) {
+          return reply(wrapError(Boom.badRequest('This Space cannot be deleted because it is reserved.')));
+        }
+
         result = await client.delete('space', id);
-      } catch(e) {
+      } catch (e) {
         return reply(wrapError(e));
       }
 
@@ -183,4 +197,20 @@ export function initSpacesApi(server) {
       pre: [routePreCheckLicenseFn]
     }
   });
+
+  async function getSpaceById(client, spaceId) {
+    try {
+      const existingSpace = await client.get('space', spaceId);
+      console.log(existingSpace);
+      return {
+        id: existingSpace.id,
+        ...existingSpace.attributes
+      };
+    } catch (e) {
+      if (client.errors.isNotFoundError(e)) {
+        return null;
+      }
+      throw e;
+    }
+  }
 }
